backend/requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
dill
supabase
python-dotenv
google-generativeai
shap
flask-cors
gunicorn
-e .
